Move frontend config.
