--- frontend/app/config/dev_conf.ts
@@ -1,7 +1,0 @@
-export class Config {
-    public dataSitekey = "6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI";
-
-    public configure() {
-        return;
-    }
-}
