+++ frontend/app/contact/contact.component.ts
@@ -3,7 +3,7 @@ import { Router } from "@angular/router";
 
 import { Contact, ContactService } from "./";
 import { CaptchaService } from "../captcha";
-import { Config } from "../config/config";
+import { Config } from "../../config/config";
 
 @Component({
     moduleId: module.id,
